Fixed a huge speed and memory issue.
I made the recursive call do the hand evaluations. That way I don't have to store the list of all k-subsets.
So now it goes way faster, and doesn't take up a bunch of memory.
Still need to clean up the code though. And I need to get this working with Polo.