I added the Reptree2 model which was built using the new training data. This tree was  built using the hand equity feature.
Because I was unable to get the program to compile from the command promt, all of my work on the interface program has been done in ecclipse. Currently the program will run and compile but it seems to break any time it receives a face card preflop or too many face cards postflop.

The following changes were made to the interface:
-Now uses handequity instead of handrank
-Speech temporarily removed for  debugging purposes.
-Immediately inputs basic game setting when run